pyproject.toml
[tool.poetry]
name = "secscanner2junit"
version = "0.1.6"
description = "Convert Security Scanner Output to JUnit Format"
authors = ["Florian Angermeir <florian.angermeir@tum.de>"]
license = "GPL-3.0-only"
readme = "README.md"
homepage = "https://github.com/angrymeir/SecScanner2JUnit"
repository = "https://github.com/angrymeir/SecScanner2JUnit"
packages = [
  { include = 'secscanner2junit' },
]

[tool.poetry.dependencies]
python = "^3.6"
junit-xml = "^1.9"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ss2ju = "secscanner2junit:main"
